requirements.txt
openai-whisper==20240314
pynput==1.7.6
PyAudio==0.2.13
python-xlib==0.35
pyyaml==6.0
